backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
gunicorn==20.1.0
whitenoise==6.4.0